Added MainViewImageView and MainViewText
Avoid duplicate code
